Update fileProcessor.ts
I have a tiny amount of experience. I want to be able to trash files with clippi but the shell object's moveItemToTrash() method/constructor IDK was removed and replaced with shell.trashItem(). The method went from synchronous to asynchronous and I am trying to learn what that means.

I am creating a new branch and proposing a change in hopes of downloading this code so I can test it to see if my implementation does anything at all. If you are reading this and have suggestions for a workflow, I am all ears. So far I have only made small classes in Python and Java so I feel a little out of my depth. 

Hopefully this is a hard to find / private / obscure commit, I am still learning github as well. Not sure what a pull request is, but here we go.